Make the network scanning code more efficient
- Only generate the IP scanning lists when needed, not every reset as
before
- Fix a bug in the UI where the enabled buttons where not in sync with
the actual robot enable state
- Fix a bug in which the communications where lost easily when the robot
address was changed
- Fix a compilation error under Visual Studio compilers
- Extend the line length to 100 characters

+++ src/MainWindow/System/CPU.cpp
@@ -34,9 +34,9 @@ static PDH_HCOUNTER cpuTotal;
 #include <QProcess>
 #endif
 
-//=============================================================================
+//==================================================================================================
 // CPU::getUsage
-//=============================================================================
+//==================================================================================================
 
 int CPU::getUsage() {
     int usage = 0;
@@ -45,7 +45,7 @@ int CPU::getUsage() {
     PDH_FMT_COUNTERVALUE counterVal;
     PdhCollectQueryData (cpuQuery);
     PdhGetFormattedCounterValue (cpuTotal, PDH_FMT_DOUBLE, 0, &counterVal);
-    usage = counterVal.doubleValue;
+    usage = (int) counterVal.doubleValue;
 #endif
 
 #if defined Q_OS_MAC
@@ -98,9 +98,9 @@ int CPU::getUsage() {
     return usage;
 }
 
-//=============================================================================
+//==================================================================================================
 // CPU::initQueryProcess
-//=============================================================================
+//==================================================================================================
 
 void CPU::initQueryProcess() {
 #if defined Q_OS_WIN
